Add the enrich_part() function + test it
