+++ kicad_parser/test_extractors_avr.py
@@ -5,10 +5,8 @@ Test suites for AVR metadata extractors.
 # pylint: disable=line-too-long
 
 import pathlib
-import yaml
 
 from extractors.avr import is_avr_datasheet, parse_pdf
-from part import Part
 
 
 def get_fixture_file(file: str) -> str:
@@ -55,81 +53,3 @@ def test_parse_pdf():
     assert pins["PA0"] == "XTAL1".split("/")
     assert pins["PA1"] == "XTAL2".split("/")
     assert pins["PA2"] == "RESET/dW".split("/")
-
-
-def test_enrich_attiny2313():
-    attiny2313_yaml = """---
-name: ATtiny2313V-10P
-description: 10MHz, 2kB Flash, 128B SRAM, 128B EEPROM, debugWIRE, DIP-20
-datasheet: http://ww1.microchip.com/downloads/en/DeviceDoc/Atmel-2543-AVR-ATtiny2313_Datasheet.pdf
-footprint: Package_DIP:DIP-20_W7.62mm
-pinout:
-  '1':
-    name: PA2/~{RESET}
-    type: bidirectional
-  '2':
-    name: PD0
-    type: bidirectional
-  '3':
-    name: PD1
-    type: bidirectional
-  '4':
-    name: PA1/XTAL2
-    type: bidirectional
-  '5':
-    name: PA0/XTAL1
-    type: bidirectional
-  '6':
-    name: PD2
-    type: bidirectional
-  '7':
-    name: PD3
-    type: bidirectional
-  '8':
-    name: PD4
-    type: bidirectional
-  '9':
-    name: PD5
-    type: bidirectional
-  '10':
-    name: GND
-    type: power_in
-  '11':
-    name: PD6
-    type: bidirectional
-  '12':
-    name: PB0
-    type: bidirectional
-  '13':
-    name: PB1
-    type: bidirectional
-  '14':
-    name: PB2
-    type: bidirectional
-  '15':
-    name: PB3
-    type: bidirectional
-  '16':
-    name: PB4
-    type: bidirectional
-  '17':
-    name: PB5
-    type: bidirectional
-  '18':
-    name: PB6
-    type: bidirectional
-  '19':
-    name: PB7
-    type: bidirectional
-  '20':
-    name: VCC
-    type: power_in
-"""
-
-    attiny2313 = Part.from_dict(yaml.safe_load(attiny2313_yaml))
-
-    # https://www.msarnoff.org/chipdb/ATtiny2313
-    print(attiny2313.pinout)
-    assert attiny2313.pinout["1"].name == "PA2/~{RESET}"
-    assert attiny2313.pinout["6"].name == "PD2"
-    assert attiny2313.pinout["20"].name == "VCC"
